Assignment2/clients/requirements.txt
flask==2.2.2
werkzeug==2.2.3
psycopg2-binary==2.9.6